+++ services/dask-sidecar/docker/boot.sh
@@ -173,7 +173,7 @@ else
     exec watchmedo auto-restart --recursive --pattern="*.py;*/src/*" --ignore-patterns="*test*;pytest_simcore/*;setup.py;*ignore*" --ignore-directories -- \
       dask worker "${DASK_SCHEDULER_URL}" \
       --local-directory /tmp/dask-sidecar \
-      --preload simcore_service_dask_sidecar.tasks \
+      --preload simcore_service_dask_sidecar.worker \
       --nworkers ${DASK_NPROCS} \
       --nthreads "${DASK_NTHREADS}" \
       --dashboard-address 8787 \
@@ -183,7 +183,7 @@ else
   else
     exec dask worker "${DASK_SCHEDULER_URL}" \
       --local-directory /tmp/dask-sidecar \
-      --preload simcore_service_dask_sidecar.tasks \
+      --preload simcore_service_dask_sidecar.worker \
       --nworkers ${DASK_NPROCS} \
       --nthreads "${DASK_NTHREADS}" \
       --dashboard-address 8787 \
