+++ services/director-v2/src/simcore_service_director_v2/modules/dask_client.py
@@ -223,7 +223,7 @@ class DaskClient:
         ) -> TaskOutputData:
             """This function is serialized by the Dask client and sent over to the Dask sidecar(s)
             Therefore, (screaming here) DO NOT MOVE THAT IMPORT ANYWHERE ELSE EVER!!"""
-            from simcore_service_dask_sidecar.tasks import (  # type: ignore[import-not-found]  # this runs inside the dask-sidecar
+            from simcore_service_dask_sidecar.worker import (  # type: ignore[import-not-found]  # this runs inside the dask-sidecar
                 run_computational_sidecar,
             )
